ALT: default: Split out the named working directory and static read-only zones
The working directory of the BIND serves for:
> any non-absolute pathnames in the configuration file are taken as
relative to this directory. The default location for most server output
files (e.g., named.run) is this directory. If a directory is not
specified, the working directory defaults to ".", the directory from
which the server was started. The directory specified should be an
absolute path, and must be writable by the effective user ID of the
named process.

This *does not* restrict the location of zone plain text files in any
way. I assume that in the past ALT BIND's maintainer wanted to
grant read-only access for such files to prevent possible harmfull
changing/adding zones by attacker in case she somehow gets an access
to filesystem with the `named` user credentials.

Actually, BIND leverages the working directory for many purposes, most
of them requires write permissions for temporary files. All of these
cases can be handled by moving out the read-only working directory
containing read-only zones into writable (by named) hierarchy while
leaving zones on read-only one.

Note: control's facility `bind-debug` blocks this change, since
it prevents the traversing of egid `named` into `/var` subpath
and thus, it has been removed.

This change greatly simplifies the maintainance of this package
otherwise *every* feature that wants named to write to some path
has to have an option to change this path. Upstream doesn't provide
such path options for everything and won't do this.

Note2: it is expected that the static zones are read-only zones.
This means that dynamic updates don't work for them since
several journal files(e.g. .jnw, .jbk) should be written to that
zone directory.