ALT: defaults: Make working directory writable
The working directory of the BIND serves for:
> any non-absolute pathnames in the configuration file are taken as
relative to this directory. The default location for most server output
files (e.g., named.run) is this directory. If a directory is not
specified, the working directory defaults to ".", the directory from
which the server was started. The directory specified should be an
absolute path, and must be writable by the effective user ID of the
named process.

Actually, `named` leverages the working directory for many purposes,
most of them requires write permissions for temporary or state files.
Upstream *requires* this to support all of its features.

Though the read-only version of installation is possible it is very
specific and limited (e.g. only authoritative server without dynamic
updates).