alt: ap: make sure CI runs over synced-to-distro repo
It's assumed that repo's BIND is latest version and always newer than
distro's one. But it was never enforced before and can lead to tests
failures or tests can be run over distro packages (not produced by
CI run).

For example,
- new BIND was pushed to distro
- CI repo was not rebased over distro one
- CI builds from its repo but installs distro BIND on container build
- tests don't run over packages built by CI but distro ones

And usually it's not quite easy to notice that.
Thereby, precheck if distro repo is ancestor of CI repo.